Split waypoint components into basic and address versions
