+++ addon/templates/components/g-map-route-address-waypoint.hbs
@@ -1,0 +1,1 @@
+{{g-map-route-waypoint routeContext lat=lat lng=lng}}
