+++ app/components/g-map-route-address-waypoint.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-g-map/components/g-map-route-address-waypoint';
