+++ tests/dummy/app/templates/application.hbs
@@ -46,7 +46,7 @@
                 originLat=37.7933 originLng=-122.4467
                 destinationLat=37.7433 destinationLng=-122.3867
                 as |routeContext|}}
-    {{g-map-route-waypoint routeContext address="Oakland, California"}}
+    {{g-map-route-address-waypoint routeContext address="Oakland, California"}}
   {{/g-map-route}}
 
   {{#g-map-address-route context
@@ -54,9 +54,9 @@
                 originAddress="Los Angeles, California"
                 destinationAddress="San Francisco, California"
                 as |routeContext|}}
-    {{g-map-route-waypoint routeContext address="Oakland, California"}}
-    {{g-map-route-waypoint routeContext address="New York City, New York"}}
-    {{g-map-route-waypoint routeContext address="Dallas, Texas"}}
+    {{g-map-route-address-waypoint routeContext address="Oakland, California"}}
+    {{g-map-route-address-waypoint routeContext address="New York City, New York"}}
+    {{g-map-route-address-waypoint routeContext address="Dallas, Texas"}}
   {{/g-map-address-route}}
 
   {{g-map-route context strokeColor=routeColor strokeWeight=6 strokeOpacity=0.5 zIndex=10
